backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.3.0
langchain-ollama>=0.2.0
langchain-google-genai>=2.0.0
langchain-text-splitters>=0.3.0
chromadb>=0.5.0
# --- PDF ---
PyMuPDF>=1.24.0
pdfplumber>=0.11.0
pypdf>=4.0.0
# --- Office ---
python-docx>=1.1.0
markitdown>=0.0.1a1
python-pptx>=0.6.23
openpyxl>=3.1.2
xlrd>=2.0.1
# --- Markup / Web ---
beautifulsoup4>=4.12.0
lxml>=5.0.0
# --- E-book / Rich text ---
ebooklib>=0.18
striprtf>=0.0.26
odfpy>=1.4.1
# --- Other ---
python-multipart>=0.0.7
